apod/pyproject.toml
[tool.poetry]
name = "apod"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
flask = "^2.2.3"
gunicorn = { extras = ["gevent"], version = "^20.1.0" }
flask-caching = "^2.0.2"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311"]

[tool.isort]
profile = "black"
